+++ flexmeasures/cli/tests/test_data_edit.py
@@ -5,12 +5,13 @@ import timely_beliefs as tb
 from sqlalchemy import select
 
 from flexmeasures.cli.tests.utils import to_flags
+from flexmeasures.data.models.audit_log import AssetAuditLog
 from flexmeasures.data.models.time_series import TimedBelief
 from flexmeasures.cli.tests.utils import get_click_commands
 from flexmeasures.tests.utils import get_test_sensor
 
 
-@pytest.mark.skip
+@pytest.mark.skip_github
 def test_add_one_sensor_attribute(app, db, setup_markets):
     from flexmeasures.cli.data_edit import edit_attribute
 
@@ -27,11 +28,48 @@ def test_add_one_sensor_attribute(app, db, setup_markets):
     result = runner.invoke(edit_attribute, to_flags(cli_input))
     assert result.exit_code == 0 and "Success" in result.output, result.exception
 
+    event = f"Updated sensor '{sensor.name}': {sensor.id} attribute 'some new attribute' to 3.0 from None"
+    assert db.session.execute(
+        select(AssetAuditLog).filter_by(
+            affected_asset_id=sensor.generic_asset_id,
+            event=event,
+            active_user_id=None,
+            active_user_name=None,
+        )
+    ).scalar_one_or_none()
+
     # Reload sensor from database and count attributes
     sensor = get_test_sensor(db)
     n_attributes_after = len(sensor.attributes)
 
     assert n_attributes_after == n_attributes_before + 1
+
+
+@pytest.mark.skip_github
+def test_update_one_asset_attribute(app, db, setup_generic_assets):
+    from flexmeasures.cli.data_edit import edit_attribute
+
+    db.session.flush()
+    asset = setup_generic_assets["test_battery"]
+    cli_input = {
+        "asset": asset.id,
+        "attribute": "some-attribute",
+        "str": "some-new-value",
+    }
+
+    runner = app.test_cli_runner()
+    result = runner.invoke(edit_attribute, to_flags(cli_input))
+    assert result.exit_code == 0 and "Success" in result.output, result.exception
+
+    event = f"Updated asset '{asset.name}': {asset.id} attribute 'some-attribute' to some-new-value from some-value"
+    assert db.session.execute(
+        select(AssetAuditLog).filter_by(
+            affected_asset_id=asset.id,
+            event=event,
+            active_user_id=None,
+            active_user_name=None,
+        )
+    ).scalar_one_or_none()
 
 
 @pytest.mark.skip_github
@@ -98,6 +136,15 @@ def test_resample_sensor_data(
     # Checksum
     assert beliefs_after["event_value"].sum() == 2 * beliefs_before["event_value"].sum()
 
+    assert db.session.execute(
+        select(AssetAuditLog).filter_by(
+            affected_asset_id=sensor.generic_asset_id,
+            event=f"Resampled sensor data for sensor '{sensor.name}': {sensor.id} to 0:30:00 from 1:00:00",
+            active_user_id=None,
+            active_user_name=None,
+        )
+    ).scalar_one_or_none()
+
     # Resample back to original resolution (on behalf of the next test case)
     cli_input["event-resolution"] = sensor.event_resolution.seconds / 60
     result = runner.invoke(
@@ -149,3 +196,14 @@ def test_transfer_ownership(app, db, add_asset_with_children, add_alternative_ac
     assert parent.owner == new_account
     for child in parent.child_assets:
         assert child.owner == new_account
+
+    for child_asset in (parent, *parent.child_assets):
+        event = f"Transfered ownership for asset '{child_asset.name}': {child_asset.id} from {old_account.id} to {new_account.id}"
+        assert db.session.execute(
+            select(AssetAuditLog).filter_by(
+                affected_asset_id=child_asset.id,
+                event=event,
+                active_user_id=None,
+                active_user_name=None,
+            )
+        ).scalar_one_or_none()